Introduce SlidingTimeWindowSample
As described in <http://taint.org/2014/01/16/145944a.html>,
exponentially-decaying reservoir is not always the right thing -
especially not with timers. One of the more correct solutions is to use
a sliding time window reservoir. The problem with them is that they can
suddenly eat a lot of memory if there are unexpectedly many updates in
the time window. With some workloads that's okay.

This patch implements such a reservoir. It's backed by a Python list
used with the `heapq` standard library module. It's probably not the
best possible data structure choice - Dropwizard Metrics seems to be
using skip lists.