Log exceptions that happen when reporting metrics
At the moment, if an exception happens inside report_now, it is wholly ignored
This means that if there is a fault in the reporter, it may silently sit there, not
reporting anything, without anyone's knowledge.

Add some basic logging of caught exceptions so that they show up in user's logs and so that
they are able to inspect problems when they arise.